# backend : offload large batches to GPU