Moves the logic of auto-offloading to the GPU when processing large batches to `ggml_backend_sched`. Currently only CUDA and Vulkan support this, this will allow any backend to support this feature. 

Instead of offloading only the matrix multiplications, the entire computation of the batch is offloaded. This reduces the amount of data that needs to be transferred between the GPU and CPU and improves performance significantly.

The weights are now copied to VRAM in the compute buffer, instead of the private CUDA pool buffer. As a result, the size of the compute buffers will increase significantly when offloading a model partially. However, the total VRAM usage should stay same, or slightly lower.

Backends that wish to support this feature need to implement the `offload_op` function. Only the CUDA backend implements it at this point.

Additionally, the CUDA backend will now attempt to register as a host pinned buffer the memory of the models, even when using mmap. Previously, host buffers were only supported with mmap disabled. This further increases the performance of automatic offloading. The usage of host pinned memory can be disabled by defining the `GGML_CUDA_NO_PINNED` environment variable.

RTX 3090 Ti, CUDA under WSL:
![bench-7b-pp1024](https://github.com/ggerganov/llama.cpp/assets/2141330/fe11f96d-93f6-4abe-b750-e716c8950e63)

![bench-mixtral-pp1024](https://github.com/ggerganov/llama.cpp/assets/2141330/402131a1-1f9d-4165-bcad-8f7ad8df5a83)

<details>
<summary>Raw data</summary>

| model                          |       size |     params | backend    | ngl |    n_batch |   n_ubatch |       mmap | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------: | ---------: | ---------- | ---------------: |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   0 |       1024 |       1024 |          0 | pp 1024    |    388.15 ± 1.24 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   0 |       1024 |       1024 |          1 | pp 1024    |    348.58 ± 2.46 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   1 |       1024 |       1024 |          0 | pp 1024    |    397.95 ± 1.48 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   1 |       1024 |       1024 |          1 | pp 1024    |    359.64 ± 1.94 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   2 |       1024 |       1024 |          0 | pp 1024    |    409.85 ± 2.36 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   2 |       1024 |       1024 |          1 | pp 1024    |    370.63 ± 3.55 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   3 |       1024 |       1024 |          0 | pp 1024    |    422.51 ± 2.51 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   3 |       1024 |       1024 |          1 | pp 1024    |    380.48 ± 1.25 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   4 |       1024 |       1024 |          0 | pp 1024    |    433.78 ± 1.37 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   4 |       1024 |       1024 |          1 | pp 1024    |    392.48 ± 2.48 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   5 |       1024 |       1024 |          0 | pp 1024    |    447.87 ± 1.24 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   5 |       1024 |       1024 |          1 | pp 1024    |    404.52 ± 2.98 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   6 |       1024 |       1024 |          0 | pp 1024    |    463.28 ± 1.98 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   6 |       1024 |       1024 |          1 | pp 1024    |    418.75 ± 2.95 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   7 |       1024 |       1024 |          0 | pp 1024    |    478.75 ± 1.40 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   7 |       1024 |       1024 |          1 | pp 1024    |    430.76 ± 2.00 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   8 |       1024 |       1024 |          0 | pp 1024    |    495.91 ± 1.76 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   8 |       1024 |       1024 |          1 | pp 1024    |    447.96 ± 3.44 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   9 |       1024 |       1024 |          0 | pp 1024    |    515.97 ± 1.26 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   9 |       1024 |       1024 |          1 | pp 1024    |    469.55 ± 2.47 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  10 |       1024 |       1024 |          0 | pp 1024    |    535.50 ± 2.02 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  10 |       1024 |       1024 |          1 | pp 1024    |    485.98 ± 3.47 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  11 |       1024 |       1024 |          0 | pp 1024    |    555.22 ± 3.80 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  11 |       1024 |       1024 |          1 | pp 1024    |    504.74 ± 3.24 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  12 |       1024 |       1024 |          0 | pp 1024    |    581.50 ± 3.47 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  12 |       1024 |       1024 |          1 | pp 1024    |    529.37 ± 1.10 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  13 |       1024 |       1024 |          0 | pp 1024    |    605.49 ± 3.89 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  13 |       1024 |       1024 |          1 | pp 1024    |    550.70 ± 1.20 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  14 |       1024 |       1024 |          0 | pp 1024    |    636.00 ± 3.93 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  14 |       1024 |       1024 |          1 | pp 1024    |    574.79 ± 1.91 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  15 |       1024 |       1024 |          0 | pp 1024    |    669.54 ± 2.16 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  15 |       1024 |       1024 |          1 | pp 1024    |    611.74 ± 3.48 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  16 |       1024 |       1024 |          0 | pp 1024    |    696.63 ± 5.25 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  16 |       1024 |       1024 |          1 | pp 1024    |    638.12 ± 2.97 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  17 |       1024 |       1024 |          0 | pp 1024    |    739.56 ± 3.48 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  17 |       1024 |       1024 |          1 | pp 1024    |    678.63 ± 3.30 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  18 |       1024 |       1024 |          0 | pp 1024    |    784.66 ± 2.44 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  18 |       1024 |       1024 |          1 | pp 1024    |    713.97 ± 2.73 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  19 |       1024 |       1024 |          0 | pp 1024    |    828.81 ± 3.28 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  19 |       1024 |       1024 |          1 | pp 1024    |    759.73 ± 3.57 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  20 |       1024 |       1024 |          0 | pp 1024    |    884.96 ± 4.69 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  20 |       1024 |       1024 |          1 | pp 1024    |    806.80 ± 6.71 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  21 |       1024 |       1024 |          0 | pp 1024    |    948.70 ± 5.85 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  21 |       1024 |       1024 |          1 | pp 1024    |    860.19 ± 5.59 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  22 |       1024 |       1024 |          0 | pp 1024    |   1019.88 ± 3.62 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  22 |       1024 |       1024 |          1 | pp 1024    |    933.79 ± 4.86 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  23 |       1024 |       1024 |          0 | pp 1024    |   1101.54 ± 4.91 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  23 |       1024 |       1024 |          1 | pp 1024    |   1007.86 ± 4.38 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  24 |       1024 |       1024 |          0 | pp 1024    |   1194.18 ± 3.30 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  24 |       1024 |       1024 |          1 | pp 1024    |  1095.93 ± 15.66 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  25 |       1024 |       1024 |          0 | pp 1024    |   1311.94 ± 8.63 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  25 |       1024 |       1024 |          1 | pp 1024    |  1207.60 ± 10.30 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  26 |       1024 |       1024 |          0 | pp 1024    |  1442.92 ± 14.07 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  26 |       1024 |       1024 |          1 | pp 1024    |  1346.63 ± 15.93 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  27 |       1024 |       1024 |          0 | pp 1024    |  1615.53 ± 15.01 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  27 |       1024 |       1024 |          1 | pp 1024    |   1490.20 ± 9.59 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  28 |       1024 |       1024 |          0 | pp 1024    |  1818.64 ± 30.18 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  28 |       1024 |       1024 |          1 | pp 1024    |  1710.29 ± 17.53 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  29 |       1024 |       1024 |          0 | pp 1024    |  2144.10 ± 21.59 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  29 |       1024 |       1024 |          1 | pp 1024    |  1993.06 ± 27.68 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  30 |       1024 |       1024 |          0 | pp 1024    |  2546.11 ± 19.09 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  30 |       1024 |       1024 |          1 | pp 1024    |  2371.35 ± 35.65 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  31 |       1024 |       1024 |          0 | pp 1024    | 2885.51 ± 115.14 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  31 |       1024 |       1024 |          1 | pp 1024    | 2863.88 ± 132.35 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  32 |       1024 |       1024 |          0 | pp 1024    | 3732.88 ± 206.19 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  32 |       1024 |       1024 |          1 | pp 1024    | 3694.50 ± 119.51 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  33 |       1024 |       1024 |          0 | pp 1024    |   4685.48 ± 5.83 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  33 |       1024 |       1024 |          1 | pp 1024    |  4653.46 ± 45.50 |

build: 4755afd1 (2431)

| model                          |       size |     params | backend    | ngl |    n_batch |   n_ubatch | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------: | ---------- | ---------------: |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   0 |       1024 |       1024 | pp 1024    |  1178.01 ± 52.80 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   1 |       1024 |       1024 | pp 1024    |  1221.25 ± 20.50 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   2 |       1024 |       1024 | pp 1024    |  1251.01 ± 30.48 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   3 |       1024 |       1024 | pp 1024    |  1294.10 ± 15.29 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   4 |       1024 |       1024 | pp 1024    |  1299.26 ± 36.69 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   5 |       1024 |       1024 | pp 1024    |  1313.64 ± 53.28 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   6 |       1024 |       1024 | pp 1024    |  1371.72 ± 48.12 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   7 |       1024 |       1024 | pp 1024    |  1404.57 ± 38.03 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   8 |       1024 |       1024 | pp 1024    |  1467.46 ± 42.10 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |   9 |       1024 |       1024 | pp 1024    |  1512.92 ± 44.17 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  10 |       1024 |       1024 | pp 1024    |  1561.79 ± 32.51 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  11 |       1024 |       1024 | pp 1024    |  1546.95 ± 33.21 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  12 |       1024 |       1024 | pp 1024    |  1638.92 ± 38.17 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  13 |       1024 |       1024 | pp 1024    |  1689.80 ± 66.00 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  14 |       1024 |       1024 | pp 1024    |  1770.98 ± 30.59 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  15 |       1024 |       1024 | pp 1024    |  1721.52 ± 79.84 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  16 |       1024 |       1024 | pp 1024    |  1806.18 ± 95.38 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  17 |       1024 |       1024 | pp 1024    |  1924.98 ± 55.63 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  18 |       1024 |       1024 | pp 1024    |  1969.87 ± 81.24 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  19 |       1024 |       1024 | pp 1024    |  2023.63 ± 63.53 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  20 |       1024 |       1024 | pp 1024    | 2105.42 ± 160.57 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  21 |       1024 |       1024 | pp 1024    | 2224.15 ± 130.01 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  22 |       1024 |       1024 | pp 1024    |  2274.62 ± 54.49 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  23 |       1024 |       1024 | pp 1024    |  2402.49 ± 98.09 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  24 |       1024 |       1024 | pp 1024    |  2598.08 ± 99.31 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  25 |       1024 |       1024 | pp 1024    |  2758.21 ± 67.71 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  26 |       1024 |       1024 | pp 1024    | 2788.94 ± 168.04 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  27 |       1024 |       1024 | pp 1024    |  3061.96 ± 81.72 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  28 |       1024 |       1024 | pp 1024    |  3219.39 ± 97.09 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  29 |       1024 |       1024 | pp 1024    |  3455.13 ± 77.40 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  30 |       1024 |       1024 | pp 1024    |  3603.32 ± 77.86 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  31 |       1024 |       1024 | pp 1024    | 3886.03 ± 106.86 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  32 |       1024 |       1024 | pp 1024    |   4449.24 ± 4.91 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CUDA       |  33 |       1024 |       1024 | pp 1024    |   4622.98 ± 8.33 |

build: 7664a45b (2441)

| model                          |       size |     params | backend    | ngl |    n_batch |   n_ubatch |       mmap | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------: | ---------: | ---------- | ---------------: |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   0 |       1024 |       1024 |          0 | pp 1024    |    134.99 ± 0.17 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   0 |       1024 |       1024 |          1 | pp 1024    |     94.98 ± 0.96 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   1 |       1024 |       1024 |          0 | pp 1024    |    137.99 ± 0.18 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   1 |       1024 |       1024 |          1 | pp 1024    |     94.75 ± 6.26 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   2 |       1024 |       1024 |          0 | pp 1024    |    140.96 ± 0.19 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   2 |       1024 |       1024 |          1 | pp 1024    |     99.24 ± 0.99 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   3 |       1024 |       1024 |          0 | pp 1024    |    144.41 ± 0.32 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   3 |       1024 |       1024 |          1 | pp 1024    |    101.79 ± 0.87 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   4 |       1024 |       1024 |          0 | pp 1024    |    147.93 ± 0.42 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   4 |       1024 |       1024 |          1 | pp 1024    |    104.54 ± 1.73 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   5 |       1024 |       1024 |          0 | pp 1024    |    151.50 ± 0.25 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   5 |       1024 |       1024 |          1 | pp 1024    |    108.43 ± 0.68 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   6 |       1024 |       1024 |          0 | pp 1024    |    155.25 ± 0.41 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   6 |       1024 |       1024 |          1 | pp 1024    |    111.04 ± 0.83 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   7 |       1024 |       1024 |          0 | pp 1024    |    159.65 ± 0.39 |
| llama 7B Q3_K - Large          |  19.03 GiB |    46.70 B | CUDA       |   7 |       1024 |       1024 |          1 | pp 1024    |    114.14 ± 1.36 |
…