@Dampfinchen 2024-03-15
Wow, I'm speechless. This is beyond incredible and a HUGE step forward!

```
llama_print_timings:        load time =   25189,16 ms
llama_print_timings:      sample time =      72,49 ms /   180 runs   (    0,40 ms per token,  2483,10 tokens per second)
llama_print_timings: prompt eval time =   31513,38 ms /  3602 tokens (    8,75 ms per token,   114,30 tokens per second)
llama_print_timings:        eval time =   41897,48 ms /   179 runs   (  234,06 ms per token,     4,27 tokens per second)
llama_print_timings:       total time =   73539,82 ms /  3781 tokens 
```
Speed before this PR:

```
llama_print_timings:        load time =    2482,92 ms
llama_print_timings:      sample time =      69,55 ms /   180 runs   (    0,39 ms per token,  2587,99 tokens per second)
llama_print_timings: prompt eval time =   51669,64 ms /  3602 tokens (   14,34 ms per token,    69,71 tokens per second)
llama_print_timings:        eval time =   42287,08 ms /   179 runs   (  236,24 ms per token,     4,23 tokens per second)
llama_print_timings:       total time =   94085,31 ms /  3781 tokens
```

That's indeed double the prompt processing speed! (5 layers offloaded with an RTX 2060 laptop.)

Thank you so much Slaren!!