@USBhost 2024-03-15
On my A6000 (using stock settings) there's a .31 tokens per second eval time regression for a 70b model. This .31 tps is consistent on just about every run.

./main -ngl 99 -m /mnt/40TB/AI/MiquMaid-v2-70B-DPO/ggml-model-Q4_K_M.gguf -p "Write a long story on why the sky is red."

```
Current HEAD
llama_print_timings:        load time =   14921.14 ms
llama_print_timings:      sample time =     405.60 ms /   723 runs   (    0.56 ms per token,  1782.53 tokens per second)
llama_print_timings: prompt eval time =     286.73 ms /    12 tokens (   23.89 ms per token,    41.85 tokens per second)
llama_print_timings:        eval time =   50749.65 ms /   722 runs   (   70.29 ms per token,    14.23 tokens per second)
llama_print_timings:       total time =   51652.08 ms /   734 tokens

llama_print_timings:        load time =   14844.16 ms
llama_print_timings:      sample time =     678.35 ms /  1187 runs   (    0.57 ms per token,  1749.83 tokens per second)
llama_print_timings: prompt eval time =     287.25 ms /    12 tokens (   23.94 ms per token,    41.78 tokens per second)
llama_print_timings:        eval time =   83862.37 ms /  1186 runs   (   70.71 ms per token,    14.14 tokens per second)
llama_print_timings:       total time =   85181.27 ms /  1198 tokens

llama_print_timings:        load time =   14820.03 ms
llama_print_timings:      sample time =     671.43 ms /  1194 runs   (    0.56 ms per token,  1778.30 tokens per second)
llama_print_timings: prompt eval time =     287.75 ms /    12 tokens (   23.98 ms per token,    41.70 tokens per second)
llama_print_timings:        eval time =   84489.73 ms /  1193 runs   (   70.82 ms per token,    14.12 tokens per second)
llama_print_timings:       total time =   85801.20 ms /  1205 tokens

PR
llama_print_timings:        load time =   16542.88 ms
llama_print_timings:      sample time =     578.93 ms /  1032 runs   (    0.56 ms per token,  1782.61 tokens per second)
llama_print_timings: prompt eval time =     287.61 ms /    12 tokens (   23.97 ms per token,    41.72 tokens per second)
llama_print_timings:        eval time =   74705.90 ms /  1031 runs   (   72.46 ms per token,    13.80 tokens per second)
llama_print_timings:       total time =   75894.40 ms /  1043 tokens

llama_print_timings:        load time =   16675.73 ms
llama_print_timings:      sample time =     476.32 ms /   831 runs   (    0.57 ms per token,  1744.63 tokens per second)
llama_print_timings: prompt eval time =     289.29 ms /    12 tokens (   24.11 ms per token,    41.48 tokens per second)
llama_print_timings:        eval time =   59736.20 ms /   830 runs   (   71.97 ms per token,    13.89 tokens per second)
llama_print_timings:       total time =   60767.06 ms /   842 tokens

llama_print_timings:        load time =   16597.13 ms
llama_print_timings:      sample time =     392.05 ms /   692 runs   (    0.57 ms per token,  1765.06 tokens per second)
llama_print_timings: prompt eval time =     292.76 ms /    12 tokens (   24.40 ms per token,    40.99 tokens per second)
llama_print_timings:        eval time =   49830.24 ms /   691 runs   (   72.11 ms per token,    13.87 tokens per second)
llama_print_timings:       total time =   50735.73 ms /   703 tokens
```