@slaren 2024-03-15
@USBhost should be fixed now.

Interestingly, this was caused by an increase to `GGML_SCHED_MAX_SPLIT_INPUTS`. Increasing this constant also increases the size of a hash table, which needs to be cleaned on every evaluation. That was enough to increase the overhead enough to be measurable.